requirements.txt
flask
werkzeug
requests
python-dotenv
pdfminer.six
pytesseract
Pillow
reportlab
